Initial Calypso commit.
C/C++ function calls work, most POD structs do as well.

Still a long way to go, but I had to upgrade to LLVM/Clang 3.5 because tail padding is added to struct initializers but not to the LLVM type in 3.4. Although Calypso isn't doing anything special it may be triggerring that strange behavior. Since CGRecordLayoutBuilder was rewritten and simplified in 3.5 it may be worth updating.