Allow inheriting from structs for foreign languages.
This is necessary for C++ because C++ classes can inherit from POD structs, and turning those into D classes if that POD struct is a base class isn't an option because you can't know in advance if other C++ headers will do that, breaking the early modules who thought this was a D struct.

Or at least I think so. In lack of a more elegant solution and since this is a kinda ugly change, this commit gathers all the changes needed to DMD and LDC to make inheriting from structs possible.