C++ class template instantiation.
Basic « from scratch » instantiation has been working for a week, the rest of the time was spent working on instantiating basic_string<> and its dependencies which involved:

- Fixing a thousand types of obscure identifier lookup issues and misc bugs.
- Managing to comprehensively support C++ partial template specializations despite the differences between C++'s and D's ones. Mapping in importAll can only be approximate, so the voodoo happens during the semantic pass.
- Groundwork for function and type alias templates.

basic_string<> is still not getting instantiated, I'm currently facing a dilemma regarding argument deduction for partial specializations. Although Clang provides those, they are already substitued to their non-dependent types, which results in identifier search failures since partial spec templare args have to be deduced during the semantic pass. I do have complex workaround in mind but can't find a simple one for now.

A note on a possible bug in leastAsSpecialized: that method is too fragile to always work with C++ templates because in :

  template<typename _Iterator,
           bool = __has_iterator_category<_Iterator>::value>
    struct __iterator_traits { };

  __iterator_traits<normal_iterator> { };

the semantic'd tiarg of the instance « __iterator_traits<normal_iterator> » will be lost and replaced by a dummy alias _Iterator = _Iterator, and then leastAsSpecialized tries to evaluate the bool expression which fails. Or is it a bug? Anyway leastAsSpecialized has to be avoided.