Check if a template instance is going to expand recursively.
See boost/mpl/aux_/integral_wrapper.hpp's "next" and "prior" constants. In C++ they are evaluated lazily while in D they lead to endless expansion of integral_c. As a workaround we try to guess during the mapping if a member of a template instance is going to cause recursive expansion and discard it if it does.