Pass the last valid SourceLocation ID to findFileRegionDecls().
Previously an arbitrary value was passed which lead to offset overflows, then 0, which caused only the first declaration of a header to be mapped. This wasn't apparent straightaway because most C headers are one big LinkageSpecDecl, and most C++ headers are one big NamespaceDecl. But this became an issue for C++ headers with declarations inside the global namespace.

It doesn't clear all the missing declaration issues yet, there's still something strange happening in qtypeinfo.h where the anonymous enum isn't found by findFileRegionDecls, but if moved further up the header it is.