Initial commit turning C++ class types from reference to value.
This removes the need of TypeValueof, which was a half-broken mess that contrived things a lot, and by harmonizing C++ POD and non-POD record semantics this solves potential ambiguities.

showcase.d is working up to the downcast from the DCXX class to the base C++ class, everything else is probably broken.