Emits functions with internal linkage too.
Previously this class was emitting functions with discardable ODR linkage, but it was really meant for functions with internal linkage.

The earliers are supposed to be emitted in any case, since the Calypso module they belong to should be imported. There are still a few instances where this doesn't happen e.g during some template instantiations. It should be further investigated later so that we can stop redundantly emitting functions with discardable ODR.