Add internal function overload alias declarations to alias a specific overload.
Not being able to do so is what forced me to make small forwarding functions in overloaded operators templates (e.g opOpAssign!"+" couldn't be mapped to operator+= directly because the later might be virtual). But now this has become a necessity because some overloaded operators return or take C++ class values which LDC doesn't handle.