Choose a SourceLocation closer to which header the declaration might "belong" and check if it's part of a Clang module.
At least for the time being since Clang 3.6's Serialization/ doesn't always map correctly the files to decls. And currently it's even varying with how the header is included, for example with:

    #include <unistd.h>

getpagesize() is mapped by ASTReader to unistd.h, but with

    #include <OGRE/Ogre.h>

where there is another extern C {} enclosing the extern C {} of cdefs.h, Clang's ASTReader maps getpagesize() to cdefs.h.

This is obviously a bug and Clang 3.7 may improve things a bit, e.g:

https://github.com/llvm-mirror/clang/commit/938847c101744983f89de82062b29814f13026f0

Modules may have to be disabled until then, now that loading the global namespace « _ » is possible (iirc).