Cast enum members values to the enum memory type.
See _ISwgraph from wctype.h: its value is 0x80000000 obtained from unsigned literals but Clang considers the enum integer type to be "int", so the implicit conversion will fail.

Another attempt to solve this was made in fromExpression() but CondExpression would cast it back to ulong because the other operand was a << operation with ulong operands, hence didn't work for _ISwgraph.