Cast operator mapping.
They are mapped to opCast templates, and similarly to overloaded operators the non-templated methods are also mapped to non-templated methods (and in this case opCast just forwards the call) which means they are overridable.

The name chosen for these non-templated methods is e.g opCast_const_bool for operator const bool() (if that's even possible). Things gets uglier for non-builtin types, where the mangled name is suffixed but this is only ever relevant to situations where one wants to override a virtual cast operator anyway, so probably fairly rarely.

From now on, they get emitted when necessary which fixes a few more linking errors.