Mapping values with anon record types is possible after all, once a few asserts are disabled.
union { ... } myUnion; doesn't have a D equivalent, but simply mapping the anonymous record as an AggregateDeclaration (not an AnonDeclaration) and setting the type of the variable to ad->type directly works, after disabling the !isAnonymous checks in StructDeclaration::semantic and ClassDeclaration::semantic.

This also fixes some ValueDecl with a typedef'd anon record type and were incorrectly skipped before.