Do not call the Clang executable anymore.
The PCH generation is now done by the Clang libs, which is the first step towards making the installation of Calypso simpler and clearer especially if there is a different system Clang executable version.

Calypso still depends on a Clang installation for builtin headers but this is less problematic and should be easy to add to CMakeLists.